Add account creation

+++ database/migrations/2020_05_01_144205_add_account_id_to_bills_table.php
@@ -14,7 +14,7 @@ class AddAccountIdToBillsTable extends Migration
     public function up()
     {
         Schema::table('bills', function (Blueprint $table) {
-            $table->unsignedBigInteger('account_id')->after('autopay');
+            $table->integer('account_id')->after('autopay');
 
             $table->foreign('account_id')->references('id')->on('accounts')->onDelete('cascade');
         });
